Move store related files into airview-store directory
