--- packages/airview-cms/src/features/airview-api/index.js
@@ -1,7 +1,0 @@
-export {
-  airviewApi,
-  useGetBranchesQuery,
-  useCreateBranchMutation,
-  useGetEntriesQuery,
-  useGetEntryQuery,
-} from "./airview-api";
