--- packages/airview-cms/src/features/airview-provider/index.js
@@ -1,1 +1,0 @@
-export { AirviewProvider } from "./airview-provider";
