+++ packages/airview-cms/src/features/airview-store/airview-provider.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Provider } from "react-redux";
-import { airviewStore } from "../airview-store";
+import { airviewStore } from "./airview-store";
 
 export function AirviewProvider({ children }) {
   return <Provider store={airviewStore}>{children}</Provider>;
