+++ packages/airview-cms/src/features/branch-creator/branch-creator.js
@@ -10,7 +10,7 @@ import {
   Typography,
 } from "@mui/material";
 import { useSelector, useDispatch } from "react-redux";
-import { useCreateBranchMutation, useGetBranchesQuery } from "../airview-api";
+import { useCreateBranchMutation, useGetBranchesQuery } from "../airview-store";
 import { selectWorkingBranch, setWorkingBranch } from "../toolbar";
 import {
   selectBranchCreatorBranchName,
