+++ packages/airview-cms/src/features/index.js
@@ -1,5 +1,3 @@
-export * from "./airview-api";
-export * from "./airview-provider";
 export * from "./airview-store";
 export * from "./branch-creator";
 export * from "./cms";
