+++ packages/airview-cms/src/features/use-get-all-entries-meta/use-get-all-entries-meta.js
@@ -1,6 +1,6 @@
 import { useSelector } from "react-redux";
 import { selectWorkingBranch } from "../toolbar";
-import { useGetBranchesQuery, useGetEntriesQuery } from "../airview-api";
+import { useGetBranchesQuery, useGetEntriesQuery } from "../airview-store";
 
 export function useGetAllEntriesMeta(select) {
   const workingBranch = useSelector(selectWorkingBranch);
